+++ tests/integration/test_actress_profile.py
@@ -17,7 +17,7 @@ from unittest.mock import patch, MagicMock
 
 @pytest.fixture(autouse=True)
 def clear_actress_cache():
-    from core.actress_scraper import _cache
+    from core.scrapers.actress.orchestrator import _cache
     _cache.clear()
     yield
     _cache.clear()
@@ -61,7 +61,7 @@ MODEL_PHP_HTML = """
 
 def test_scrape_graphis_photo_success():
     """測試 Graphis 爬蟲成功案例（Mock）— 雙 request: listing + model.php"""
-    from core.graphis_scraper import scrape_graphis_photo
+    from core.scrapers.actress.graphis import scrape_graphis_photo
 
     with patch('requests.get') as mock_get:
         listing_resp = MagicMock(status_code=200, text=_LISTING_HTML)
@@ -86,7 +86,7 @@ def test_scrape_graphis_photo_success():
 
 def test_scrape_graphis_photo_not_found():
     """測試女優不存在於 Graphis（Mock）"""
-    from core.graphis_scraper import scrape_graphis_photo
+    from core.scrapers.actress.graphis import scrape_graphis_photo
 
     mock_html = """
     <html>
@@ -113,7 +113,7 @@ def test_scrape_graphis_photo_not_found():
 
 def test_scrape_graphis_photo_timeout():
     """測試 timeout 處理（Mock）"""
-    from core.graphis_scraper import scrape_graphis_photo
+    from core.scrapers.actress.graphis import scrape_graphis_photo
     import requests
 
     with patch('requests.get') as mock_get:
@@ -125,7 +125,7 @@ def test_scrape_graphis_photo_timeout():
 
 def test_scrape_graphis_photo_request_error():
     """測試 request error 處理（Mock）"""
-    from core.graphis_scraper import scrape_graphis_photo
+    from core.scrapers.actress.graphis import scrape_graphis_photo
     import requests
 
     with patch('requests.get') as mock_get:
@@ -141,7 +141,7 @@ def test_scrape_graphis_photo_request_error():
 
 def test_get_actress_profile_both_sources():
     """測試雙來源合併邏輯（Mock）"""
-    from core.actress_scraper import get_actress_profile, _cache
+    from core.scrapers.actress.orchestrator import get_actress_profile, _cache
 
     # Mock Graphis 回傳照片
     def mock_graphis(name):
@@ -162,9 +162,9 @@ def test_get_actress_profile_both_sources():
             'cup': 'G'
         }
 
-    with patch('core.graphis_scraper.scrape_graphis_photo', side_effect=mock_graphis), \
-         patch('core.actress_scraper.scrape_actress_profile', side_effect=mock_javbus), \
-         patch('core.gfriends_lookup.lookup_gfriends', return_value=None):
+    with patch('core.scrapers.actress.graphis.scrape_graphis_photo', side_effect=mock_graphis), \
+         patch('core.scrapers.actress.javbus.scrape_actress_profile', side_effect=mock_javbus), \
+         patch('core.scrapers.actress.gfriends.lookup_gfriends', return_value=None):
 
         result = get_actress_profile("桜空もも")
 
@@ -182,7 +182,7 @@ def test_get_actress_profile_both_sources():
 
 def test_get_actress_profile_javbus_only():
     """測試只有 JavBus 有資料（Mock）"""
-    from core.actress_scraper import get_actress_profile, _cache
+    from core.scrapers.actress.orchestrator import get_actress_profile, _cache
 
     def mock_javbus(name):
         return {
@@ -192,9 +192,9 @@ def test_get_actress_profile_javbus_only():
             'age': 29
         }
 
-    with patch('core.graphis_scraper.scrape_graphis_photo', return_value=None), \
-         patch('core.actress_scraper.scrape_actress_profile', side_effect=mock_javbus), \
-         patch('core.gfriends_lookup.lookup_gfriends', return_value=None):
+    with patch('core.scrapers.actress.graphis.scrape_graphis_photo', return_value=None), \
+         patch('core.scrapers.actress.javbus.scrape_actress_profile', side_effect=mock_javbus), \
+         patch('core.scrapers.actress.gfriends.lookup_gfriends', return_value=None):
 
         result = get_actress_profile("桜空もも")
 
@@ -205,7 +205,7 @@ def test_get_actress_profile_javbus_only():
 
 def test_get_actress_profile_graphis_only():
     """測試只有 Graphis 有資料（Mock）"""
-    from core.actress_scraper import get_actress_profile, _cache
+    from core.scrapers.actress.orchestrator import get_actress_profile, _cache
 
     def mock_graphis(name):
         return {
@@ -214,9 +214,9 @@ def test_get_actress_profile_graphis_only():
             'backdrop_url': 'https://graphis.ne.jp/model.jpg'
         }
 
-    with patch('core.graphis_scraper.scrape_graphis_photo', side_effect=mock_graphis), \
-         patch('core.actress_scraper.scrape_actress_profile', return_value=None), \
-         patch('core.gfriends_lookup.lookup_gfriends', return_value=None):
+    with patch('core.scrapers.actress.graphis.scrape_graphis_photo', side_effect=mock_graphis), \
+         patch('core.scrapers.actress.javbus.scrape_actress_profile', return_value=None), \
+         patch('core.scrapers.actress.gfriends.lookup_gfriends', return_value=None):
 
         result = get_actress_profile("桜空もも")
 
@@ -228,10 +228,10 @@ def test_get_actress_profile_graphis_only():
 
 def test_get_actress_profile_both_fail():
     """測試雙來源都失敗（Mock）"""
-    from core.actress_scraper import get_actress_profile, _cache
+    from core.scrapers.actress.orchestrator import get_actress_profile, _cache
 
-    with patch('core.graphis_scraper.scrape_graphis_photo', return_value=None), \
-         patch('core.actress_scraper.scrape_actress_profile', return_value=None):
+    with patch('core.scrapers.actress.graphis.scrape_graphis_photo', return_value=None), \
+         patch('core.scrapers.actress.javbus.scrape_actress_profile', return_value=None):
 
         result = get_actress_profile("不存在的女優")
         assert result is None
@@ -243,7 +243,7 @@ def test_get_actress_profile_both_fail():
 
 def test_get_actress_profile_cache_hit():
     """測試 cache 命中（Mock）"""
-    from core.actress_scraper import get_actress_profile, _cache
+    from core.scrapers.actress.orchestrator import get_actress_profile, _cache
 
     def mock_graphis(name):
         return {
@@ -260,9 +260,9 @@ def test_get_actress_profile_cache_hit():
             'age': 29
         }
 
-    with patch('core.graphis_scraper.scrape_graphis_photo', side_effect=mock_graphis) as graphis_mock, \
-         patch('core.actress_scraper.scrape_actress_profile', side_effect=mock_javbus) as javbus_mock, \
-         patch('core.gfriends_lookup.lookup_gfriends', return_value=None):
+    with patch('core.scrapers.actress.graphis.scrape_graphis_photo', side_effect=mock_graphis) as graphis_mock, \
+         patch('core.scrapers.actress.javbus.scrape_actress_profile', side_effect=mock_javbus) as javbus_mock, \
+         patch('core.scrapers.actress.gfriends.lookup_gfriends', return_value=None):
 
         # 第一次呼叫
         result1 = get_actress_profile("桜空もも")
@@ -281,7 +281,7 @@ def test_get_actress_profile_cache_hit():
 
 def test_get_actress_profile_cache_expired():
     """測試 cache 過期（Mock）"""
-    from core.actress_scraper import get_actress_profile, _cache, _CACHE_TTL
+    from core.scrapers.actress.orchestrator import get_actress_profile, _cache, _CACHE_TTL
 
     def mock_graphis(name):
         return {
@@ -303,9 +303,9 @@ def test_get_actress_profile_cache_expired():
         return fixed_time[0]
 
     with patch('time.time', side_effect=mock_time), \
-         patch('core.graphis_scraper.scrape_graphis_photo', side_effect=mock_graphis) as graphis_mock, \
-         patch('core.actress_scraper.scrape_actress_profile', side_effect=mock_javbus) as javbus_mock, \
-         patch('core.gfriends_lookup.lookup_gfriends', return_value=None):
+         patch('core.scrapers.actress.graphis.scrape_graphis_photo', side_effect=mock_graphis) as graphis_mock, \
+         patch('core.scrapers.actress.javbus.scrape_actress_profile', side_effect=mock_javbus) as javbus_mock, \
+         patch('core.scrapers.actress.gfriends.lookup_gfriends', return_value=None):
 
         # 第一次呼叫
         result1 = get_actress_profile("桜空もも")
@@ -323,7 +323,7 @@ def test_get_actress_profile_cache_expired():
 
 def test_get_actress_profile_cache_name_normalization():
     """測試 cache key 名稱正規化（Mock）"""
-    from core.actress_scraper import get_actress_profile, _cache
+    from core.scrapers.actress.orchestrator import get_actress_profile, _cache
 
     def mock_graphis(name):
         return {
@@ -335,9 +335,9 @@ def test_get_actress_profile_cache_name_normalization():
     def mock_javbus(name):
         return {'name': name, 'img': 'https://javbus.com/img.jpg'}
 
-    with patch('core.graphis_scraper.scrape_graphis_photo', side_effect=mock_graphis) as graphis_mock, \
-         patch('core.actress_scraper.scrape_actress_profile', side_effect=mock_javbus), \
-         patch('core.gfriends_lookup.lookup_gfriends', return_value=None):
+    with patch('core.scrapers.actress.graphis.scrape_graphis_photo', side_effect=mock_graphis) as graphis_mock, \
+         patch('core.scrapers.actress.javbus.scrape_actress_profile', side_effect=mock_javbus), \
+         patch('core.scrapers.actress.gfriends.lookup_gfriends', return_value=None):
 
         # 第一次：正常名稱
         result1 = get_actress_profile("桜空もも")
@@ -494,7 +494,7 @@ def mock_actress_profile():
 def test_search_api_actress_with_profile(client, mock_search_actress, mock_actress_profile):
     """測試女優搜尋 API（有 actress_profile）"""
     with patch('web.routers.search.search_actress', side_effect=mock_search_actress), \
-         patch('core.actress_scraper.get_actress_profile', side_effect=mock_actress_profile):
+         patch('core.scrapers.actress.orchestrator.get_actress_profile', side_effect=mock_actress_profile):
 
         resp = client.get("/api/search?q=桜空もも&mode=actress")
         data = resp.json()
@@ -573,7 +573,7 @@ def test_search_api_graceful_failure(client, mock_search_actress):
     """測試雙來源失敗時不影響搜尋結果"""
     # Mock 雙來源都失敗
     with patch('web.routers.search.search_actress', side_effect=mock_search_actress), \
-         patch('core.actress_scraper.get_actress_profile', return_value=None):
+         patch('core.scrapers.actress.orchestrator.get_actress_profile', return_value=None):
 
         resp = client.get("/api/search?q=桜空もも&mode=actress")
         data = resp.json()
@@ -601,7 +601,7 @@ def test_search_api_variant_id_no_profile(client):
 def test_sse_includes_actress_profile(client, mock_search_actress, mock_actress_profile):
     """SSE result event 應包含 actress_profile"""
     with patch('web.routers.search.smart_search', side_effect=mock_search_actress), \
-         patch('core.actress_scraper.get_actress_profile', side_effect=mock_actress_profile):
+         patch('core.scrapers.actress.orchestrator.get_actress_profile', side_effect=mock_actress_profile):
 
         response = client.get('/api/search/stream?q=桜空もも')
 
@@ -630,7 +630,7 @@ def test_sse_includes_actress_profile(client, mock_search_actress, mock_actress_
 
 def test_lookup_gfriends_top1_hit():
     """Top1 片商 folder HEAD 200 → 回傳 CDN URL"""
-    from core.gfriends_lookup import lookup_gfriends, CDN_BASE
+    from core.scrapers.actress.gfriends import lookup_gfriends, CDN_BASE
 
     with patch('requests.head') as mock_head:
         hit_resp = MagicMock(status_code=200)
@@ -646,7 +646,7 @@ def test_lookup_gfriends_top1_hit():
 
 def test_lookup_gfriends_top1_miss_top2_hit():
     """Top1 非 200, Top2 命中 → fallback 成功"""
-    from core.gfriends_lookup import lookup_gfriends, CDN_BASE
+    from core.scrapers.actress.gfriends import lookup_gfriends, CDN_BASE
 
     with patch('requests.head') as mock_head:
         miss_resp = MagicMock(status_code=403)
@@ -665,7 +665,7 @@ def test_lookup_gfriends_top1_miss_top2_hit():
 
 def test_lookup_gfriends_avdbs_fallback():
     """Top1 + Top2 folder 都 miss → AVDBS fallback 命中"""
-    from core.gfriends_lookup import lookup_gfriends, CDN_BASE, FALLBACK_FOLDER
+    from core.scrapers.actress.gfriends import lookup_gfriends, CDN_BASE, FALLBACK_FOLDER
 
     with patch('requests.head') as mock_head:
         miss_resp = MagicMock(status_code=404)
@@ -689,7 +689,7 @@ def test_lookup_gfriends_avdbs_fallback():
 
 def test_lookup_gfriends_all_miss():
     """全部 folder 都非 200 → 回傳 None"""
-    from core.gfriends_lookup import lookup_gfriends
+    from core.scrapers.actress.gfriends import lookup_gfriends
 
     with patch('requests.head') as mock_head:
         miss_resp = MagicMock(status_code=404)
@@ -707,7 +707,7 @@ def test_lookup_gfriends_all_miss():
 
 def test_lookup_gfriends_ai_fix_prefix():
     """{name}.jpg 非 200，AI-Fix-{name}.jpg 命中 → 回傳 AI-Fix URL"""
-    from core.gfriends_lookup import lookup_gfriends, CDN_BASE
+    from core.scrapers.actress.gfriends import lookup_gfriends, CDN_BASE
 
     with patch('requests.head') as mock_head:
         miss_resp = MagicMock(status_code=403)
@@ -724,7 +724,7 @@ def test_lookup_gfriends_ai_fix_prefix():
 
 def test_lookup_gfriends_timeout():
     """requests.head 拋出 Timeout → 回傳 None（fail-open）"""
-    from core.gfriends_lookup import lookup_gfriends
+    from core.scrapers.actress.gfriends import lookup_gfriends
     import requests
 
     with patch('requests.head') as mock_head:
@@ -741,7 +741,7 @@ def test_lookup_gfriends_timeout():
 
 def test_parse_graphis_profile_full():
     """完整 model.php HTML → 所有欄位正確解析"""
-    from core.graphis_scraper import _parse_graphis_profile
+    from core.scrapers.actress.graphis import _parse_graphis_profile
 
     result = _parse_graphis_profile(MODEL_PHP_HTML)
 
@@ -757,7 +757,7 @@ def test_parse_graphis_profile_full():
 
 def test_parse_graphis_profile_bwh_regex():
     """BWH regex 解析：有 cup 與無 cup 兩種情境"""
-    from core.graphis_scraper import _parse_graphis_profile
+    from core.scrapers.actress.graphis import _parse_graphis_profile
 
     # With cup: B86(E) W61 H88
     html_with_cup = """
@@ -795,7 +795,7 @@ def test_parse_graphis_profile_bwh_regex():
 
 def test_scrape_graphis_photo_with_profile():
     """listing + model.php 雙 mock → profile 欄位存在於回傳值"""
-    from core.graphis_scraper import scrape_graphis_photo
+    from core.scrapers.actress.graphis import scrape_graphis_photo
 
     with patch('requests.get') as mock_get:
         listing_resp = MagicMock(status_code=200, text=_LISTING_HTML)
@@ -819,7 +819,7 @@ def test_scrape_graphis_photo_with_profile():
 
 def test_scrape_graphis_photo_profile_fail():
     """model.php 回傳 500 → 仍回傳圖片 URL，profile 欄位為空"""
-    from core.graphis_scraper import scrape_graphis_photo
+    from core.scrapers.actress.graphis import scrape_graphis_photo
 
     with patch('requests.get') as mock_get:
         listing_resp = MagicMock(status_code=200, text=_LISTING_HTML)
@@ -883,13 +883,13 @@ def _make_javbus_result(name="桜空もも", **overrides):
 
 def test_get_actress_profile_gfriends_wins():
     """gfriends 圖片優先級最高：img 應為 gfriends URL，而非 graphis/javbus"""
-    from core.actress_scraper import get_actress_profile, _cache
+    from core.scrapers.actress.orchestrator import get_actress_profile, _cache
 
     gfriends_url = 'https://cdn.jsdelivr.net/gh/gfriends/gfriends@master/Content/7-S1/桜空もも.jpg'
 
-    with patch('core.graphis_scraper.scrape_graphis_photo', return_value=_make_graphis_result()), \
-         patch('core.actress_scraper.scrape_actress_profile', return_value=_make_javbus_result()), \
-         patch('core.gfriends_lookup.lookup_gfriends', return_value=gfriends_url):
+    with patch('core.scrapers.actress.graphis.scrape_graphis_photo', return_value=_make_graphis_result()), \
+         patch('core.scrapers.actress.javbus.scrape_actress_profile', return_value=_make_javbus_result()), \
+         patch('core.scrapers.actress.gfriends.lookup_gfriends', return_value=gfriends_url):
 
         result = get_actress_profile("桜空もも", makers=['S1'])
 
@@ -902,11 +902,11 @@ def test_get_actress_profile_gfriends_wins():
 
 def test_get_actress_profile_graphis_text_wins():
     """graphis 文字欄位優先於 javbus（age/height/cup）"""
-    from core.actress_scraper import get_actress_profile, _cache
+    from core.scrapers.actress.orchestrator import get_actress_profile, _cache
 
-    with patch('core.graphis_scraper.scrape_graphis_photo', return_value=_make_graphis_result(age=28, height='160cm', cup='G')), \
-         patch('core.actress_scraper.scrape_actress_profile', return_value=_make_javbus_result(age=27, height='155cm', cup='F')), \
-         patch('core.gfriends_lookup.lookup_gfriends', return_value=None):
+    with patch('core.scrapers.actress.graphis.scrape_graphis_photo', return_value=_make_graphis_result(age=28, height='160cm', cup='G')), \
+         patch('core.scrapers.actress.javbus.scrape_actress_profile', return_value=_make_javbus_result(age=27, height='155cm', cup='F')), \
+         patch('core.scrapers.actress.gfriends.lookup_gfriends', return_value=None):
 
         result = get_actress_profile("桜空もも")
 
@@ -919,11 +919,11 @@ def test_get_actress_profile_graphis_text_wins():
 
 def test_get_actress_profile_name_en():
     """graphis name_en 正確傳遞到最終結果"""
-    from core.actress_scraper import get_actress_profile, _cache
+    from core.scrapers.actress.orchestrator import get_actress_profile, _cache
 
-    with patch('core.graphis_scraper.scrape_graphis_photo', return_value=_make_graphis_result(name_en='Momo Sakurazora')), \
-         patch('core.actress_scraper.scrape_actress_profile', return_value=_make_javbus_result()), \
-         patch('core.gfriends_lookup.lookup_gfriends', return_value=None):
+    with patch('core.scrapers.actress.graphis.scrape_graphis_photo', return_value=_make_graphis_result(name_en='Momo Sakurazora')), \
+         patch('core.scrapers.actress.javbus.scrape_actress_profile', return_value=_make_javbus_result()), \
+         patch('core.scrapers.actress.gfriends.lookup_gfriends', return_value=None):
 
         result = get_actress_profile("桜空もも")
 
@@ -933,7 +933,7 @@ def test_get_actress_profile_name_en():
 
 def test_get_actress_profile_birth_javbus_only():
     """birth/hometown 只來自 javbus，不被 graphis 覆蓋"""
-    from core.actress_scraper import get_actress_profile, _cache
+    from core.scrapers.actress.orchestrator import get_actress_profile, _cache
 
     javbus_birth = '1996-12-03'
     javbus_hometown = '東京都'
@@ -943,9 +943,9 @@ def test_get_actress_profile_birth_javbus_only():
 
     javbus = _make_javbus_result(birth=javbus_birth, hometown=javbus_hometown)
 
-    with patch('core.graphis_scraper.scrape_graphis_photo', return_value=graphis), \
-         patch('core.actress_scraper.scrape_actress_profile', return_value=javbus), \
-         patch('core.gfriends_lookup.lookup_gfriends', return_value=None):
+    with patch('core.scrapers.actress.graphis.scrape_graphis_photo', return_value=graphis), \
+         patch('core.scrapers.actress.javbus.scrape_actress_profile', return_value=javbus), \
+         patch('core.scrapers.actress.gfriends.lookup_gfriends', return_value=None):
 
         result = get_actress_profile("桜空もも")
 
@@ -957,13 +957,13 @@ def test_get_actress_profile_birth_javbus_only():
 
 def test_get_actress_profile_gfriends_only():
     """gfriends 有圖但 javbus/graphis 都沒資料 → 仍回傳 minimal profile"""
-    from core.actress_scraper import get_actress_profile, _cache
+    from core.scrapers.actress.orchestrator import get_actress_profile, _cache
 
     gfriends_url = 'https://cdn.jsdelivr.net/gh/gfriends/gfriends@master/Content/7-S1/桜空もも.jpg'
 
-    with patch('core.graphis_scraper.scrape_graphis_photo', return_value=None), \
-         patch('core.actress_scraper.scrape_actress_profile', return_value=None), \
-         patch('core.gfriends_lookup.lookup_gfriends', return_value=gfriends_url):
+    with patch('core.scrapers.actress.graphis.scrape_graphis_photo', return_value=None), \
+         patch('core.scrapers.actress.javbus.scrape_actress_profile', return_value=None), \
+         patch('core.scrapers.actress.gfriends.lookup_gfriends', return_value=gfriends_url):
 
         result = get_actress_profile("桜空もも", makers=['S1'])
 
@@ -993,7 +993,7 @@ def test_search_api_passes_makers_to_profile(client):
 
     # Patch smart_search in the router's own namespace (it's imported at module load time)
     with patch('web.routers.search.smart_search', side_effect=mock_smart_search), \
-         patch('core.actress_scraper.get_actress_profile', mock_profile):
+         patch('core.scrapers.actress.orchestrator.get_actress_profile', mock_profile):
 
         resp = client.get("/api/search?q=桜空もも")
         data = resp.json()
@@ -1026,7 +1026,7 @@ def test_sse_passes_makers_to_profile(client):
     })
 
     with patch('web.routers.search.smart_search', side_effect=mock_smart_search), \
-         patch('core.actress_scraper.get_actress_profile', mock_profile):
+         patch('core.scrapers.actress.orchestrator.get_actress_profile', mock_profile):
 
         response = client.get('/api/search/stream?q=桜空もも')
